resolve crash: malloc(-1024);

+++ includes/zone.h
@@ -32,7 +32,7 @@ enum					e_zone_type
 struct					s_zone_meta
 {
 	enum e_zone_type	type;
-	unsigned int		size;
+	size_t				size;
 	struct s_zone_meta	*next;
 };
 
@@ -40,7 +40,7 @@ typedef struct s_zone_meta	t_zone;
 # define ZONE_TO_BLOCK(zone)	((t_block *)((char *)zone + sizeof(t_zone)))
 # define BLOCK_TO_ZONE(block)	((t_zone *)((char *)block - sizeof(t_zone)))
 
-struct s_zone_meta		*mmap_zone(unsigned long size);
+struct s_zone_meta		*mmap_zone(size_t size);
 void					insert_zone_to_list(struct s_zone_meta *zone);
 enum e_zone_type		define_zone_type(size_t size);
 struct s_zone_meta		*get_my_zone_meta(struct s_block_meta *block);
